test: 리스폰 뷰모델 테스트 추가 및 공통 mock 분리
- requestRespawn 동작/중복/게임종료 무시 케이스 테스트 추가
- 테스트용 Mock을 Support/TestDoubles.swift로 분리해 재사용

refs dev-97
